[Essential]Move LocalFeedLoader and FeedStore protocol to its own production file
